+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- `project-name` now has a `--unity-exclusion` flag which sets the exclusion field to default more appropriate for Unity.
+
 ### Removed
 
 ## [2.4.0] 2023-11-15